+++ compliance-snapshot/app/core/config.py
@@ -1,0 +1,1 @@
+# Configuration settings
