+++ compliance-snapshot/app/core/utils.py
@@ -1,0 +1,1 @@
+# Utility functions
