+++ compliance-snapshot/app/templates/upload.html
@@ -1,0 +1,1 @@
+<!-- Upload form template -->
