+++ compliance-snapshot/requirements.txt
@@ -1,0 +1,9 @@
+fastapi==0.111.*
+uvicorn[standard]==0.30.*
+python-multipart==0.0.9
+pandas
+openpyxl
+matplotlib
+reportlab
+jinja2
+pillow
